The ID in the database is now just a copy of utc_time, hence sorting by ID produces the desired results. The other way round was to remove ID, and change  the references to it to utc_time.
The mergeddata creation is done in one go rather than via a scratchdata view
( just seemed neater ).

There is an option to send the string 'NOW ' rather than a time, and the Tank
will replace that with its own time.

Every time a new flight_num on CORCON is detected the database is emptied.

If when you start the listener the last time is more than 1 hour ago
- it clears the data.

There are a couple alternatives for the time_past_midnight. To calulate
midnight they either use the earliest time in the database ( retrieved every
time ) or calculated once when the decades_server starts.